third_party/WebKit/LayoutTests/editing/pasteboard/paste-text-events-expected.txt

This tests that Paste commands fires textInput events.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS event.data is 'PlainHello'
PASS testTargetTextarea.value is 'PlainHello'
PASS event.data is 'PlainHello'
PASS testTargetInput.value is 'PlainHello'
PASS event.data is ''
PASS testTargetEditable.innerHTML is 'PlainHello'
PASS event.data is 'RichHello'
PASS testTargetTextarea.value is 'RichHello'
PASS event.data is 'RichHello'
PASS testTargetInput.value is 'RichHello'
PASS event.data is ''
PASS testTargetEditable.innerHTML is '<b>Rich</b>Hello'
PASS textInputCount is proceedingTestCases.length
PASS event.data is 'PlainHello'
PASS testTargetTextarea.value is ''
PASS event.data is 'PlainHello'
PASS testTargetInput.value is ''
PASS event.data is ''
PASS testTargetEditable.innerHTML is ''
PASS event.data is 'RichHello'
PASS testTargetTextarea.value is ''
PASS event.data is 'RichHello'
PASS testTargetInput.value is ''
PASS event.data is ''
PASS testTargetEditable.innerHTML is ''
PASS textInputCount is cancelingTestCases.length
PASS successfullyParsed is true

TEST COMPLETE
